fix all debts command
