+++ App/Commands/AlldebtsCommand.php
@@ -1,5 +1,6 @@
 <?php
 
+
 namespace Longman\TelegramBot\Commands\UserCommands;
 
 use App\Model\DebtTable;
@@ -9,10 +10,13 @@ use Longman\TelegramBot\Request;
 
 class AlldebtsCommand extends UserCommand
 {
-    protected $name = 'debt';                                 // Your command's name
-    protected $description = 'get all debts for asking user'; // Your command description
-    protected $usage = '/my_debt';                               // Usage of your command
-    protected $version = '1.0.0';                             // Version of your command
+    // TODO почему-то не работает - разобраться у меня не получилось.
+    //  Но вроде эта команда не очень нужна, т.к. /calculate нормально работает
+    
+    protected $name = 'all_debts';                                 // Your command's name
+    protected $description = 'get all debts for session';          // Your command description
+    protected $usage = '/alldebts';                               // Usage of your command
+    protected $version = '1.0.0';                                  // Version of your command
 
     public function execute()
     {
@@ -20,24 +24,40 @@ class AlldebtsCommand extends UserCommand
 
         $chatId = $message->getChat()->getId();   // Get the current Chat ID
 
-        $user = $message->getFrom()->getUsername();
-
         $message->getChat()->getUsername();
         $sessionId = (new SessionTable())->getLastActiveSessionByChatId($chatId);
-        $debtsData = (new DebtTable())->getAllActiveDebts($sessionId);
-
-        $debtText = 'Таблица долгов на текущий момент:' . PHP_EOL;
-        foreach ($debtsData as $debt)
-        {
-            $debtText .= "{$debt['user_debtor']} должен {$debt['user_creditor']} {$debt['amount']} за \"{$debt['description']}\".\n";
-        }
 
         $data = [                                  // Set up the new message data
             'chat_id' => $chatId,                  // Set Chat ID to send the message to
-            'text'    => $debtText                 // Set message to send
+            'text'    => $this->prepareRawDebtsText($sessionId)                 // Set message to send
         ];
 
 
         return Request::sendMessage($data);        // Send message!
+    }
+
+    /*
+     *  Все долги по текущей сессии без аггрегации, но с комментариями
+     */
+    private function prepareRawDebtsText(int $session)
+    {
+        $headerText = "Таблица долгов на текущий момент: \n";
+        $debtText = '';
+        $debtsData = $this->getDebtTable()->getAllActiveDebts($session);
+        foreach ($debtsData as $debt) {
+            if (isset($debt['user_debtor']) && isset($debt['user_creditor'])
+                && isset($debt['amount']) && isset($debt['description'])) {
+                $debtText .= "{$debt['user_debtor']} должен {$debt['user_creditor']} {$debt['amount']} за \"{$debt['description']}\".\n";
+            }
+        }
+        if (empty($debtText)) {
+            $debtText =  "Поздравляю! У вас нет долгов в текущей сессии";
+        }
+        return $headerText . $debtText;
+    }
+
+    private function getDebtTable()
+    {
+        return new DebtTable();
     }
 }
